+++ lib/ext/Event2.dart
@@ -130,6 +130,7 @@ String? event2TypeFilterToDisplayString(Event2TypeFilter? value) {
     case Event2TypeFilter.paid: return Localization().getStringEx('model.event2.event_type.paid', 'Paid');
     case Event2TypeFilter.inPerson: return Localization().getStringEx('model.event2.event_type.in_person', 'In-person');
     case Event2TypeFilter.online: return Localization().getStringEx('model.event2.event_type.online', 'Online');
+    case Event2TypeFilter.hybrid: return Localization().getStringEx('model.event2.event_type.hybrid', 'Hybrid');
     case Event2TypeFilter.public: return Localization().getStringEx('model.event2.event_type.public', 'Public');
     case Event2TypeFilter.private: return Localization().getStringEx('model.event2.event_type.private', 'Private');
     case Event2TypeFilter.nearby: return Localization().getStringEx('model.event2.event_type.nearby', 'Nearby');