+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Hooked up first cropped version of Create Event2 panel [#3300](https://github.com/rokwire/illinois-app/issues/3300).
 
 ## [5.0.15] - 2023-06-26
 ### Added